+++ builder_engine/frontend/index.js
@@ -1,5 +1,5 @@
 export * from "./app/index.js";
 export * from "./document/index.js";
 export * from "./sdk/index.js";
-
 export * from "./canvas/index.js";
+export * from "./inspector/index.js";